+++ tests/testthat/test-cdm.R
@@ -11,9 +11,13 @@ test_that("cdm reference works locally", {
 
   expect_true(is.character(listTables(con, schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"))))
 
-  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
 
-  expect_true(is.null(verify_write_access(con, write_schema = "scratch")))
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
+
+  expect_true(is.null(CDMConnector:::verify_write_access(con, write_schema = "scratch")))
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -35,6 +39,9 @@ test_that("cdm reference works on postgres", {
   expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"))))
 
   cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+
+  expect_error(assert_tables(cdm, "person"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -60,6 +67,9 @@ test_that("cdm reference works on sql server", {
 
   cdm <- cdm_from_con(con, cdm_schema = c("CDMV5", "dbo"), cdm_tables = tbl_group("vocab"))
 
+  expect_error(assert_tables(cdm, "person"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
 
@@ -67,9 +77,9 @@ test_that("cdm reference works on sql server", {
   expect_null(verify_write_access(con, write_schema = c("tempdb", "dbo")))
 
   cohort <- dplyr::tibble(cohort_id = 1L,
-                           subject_id = 1L:2L,
-                           cohort_start_date = c(Sys.Date(), as.Date("2020-02-03")),
-                           cohort_end_date = c(Sys.Date(), as.Date("2020-11-04")))
+                          subject_id = 1L:2L,
+                          cohort_start_date = c(Sys.Date(), as.Date("2020-02-03")),
+                          cohort_end_date = c(Sys.Date(), as.Date("2020-11-04")))
 
   DBI::dbWriteTable(con, DBI::Id(catalog = "tempdb", schema = "dbo", table = "cohort"), cohort, overwrite = TRUE)
 
@@ -103,7 +113,10 @@ test_that("cdm reference works on redshift", {
 
   expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"))))
 
-  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -121,7 +134,11 @@ test_that("cdm reference works on duckdb", {
 
   expect_true(is.character(listTables(con)))
 
-  cdm <- cdm_from_con(con, cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -178,7 +195,7 @@ test_that("collect a cdm", {
 test_that("stow and cdm_from_files works", {
   save_path <- file.path(tempdir(), paste0("tmp_", paste(sample(letters, 10, replace = TRUE), collapse = "")))
   dir.create(save_path)
-  cdm_tables <- c("person", "observation_period")
+  cdm_tables <- c("person", "observation_period", "cdm_source", "vocabulary")
   con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
 
   # Test tidyselect in cdm_from_con. Should not produce message about ambiguous names.
@@ -192,23 +209,20 @@ test_that("stow and cdm_from_files works", {
 
   stow(cdm, path = save_path)
 
-  expect_equal(list.files(save_path), c("observation_period.parquet" ,"person.parquet"))
+  expect_setequal(list.files(save_path), paste0(cdm_tables, ".parquet"))
 
-  # test tidyselect in cdm_from_files. Should not produce message about ambiguous names.
-  expect_message(cdm_from_files(save_path, cdm_tables = matches("person|observation_period")), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = c(person, observation_period)), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = c("person", "observation_period")), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = all_of(cdm_tables)), NA)
+  expect_message(cdm_from_files(save_path), NA)
+  expect_warning(cdm_from_files(save_path, cdm_tables = all_of(cdm_tables)), "deprecated")
 
-  local_cdm <- cdm_from_files(save_path, cdm_tables = all_of(cdm_tables))
+  local_cdm <- cdm_from_files(save_path)
   expect_s3_class(local_cdm, "cdm_reference")
   expect_equal(local_cdm$person, collect(cdm$person))
-  expect_output(validate_cdm(local_cdm))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
 
-  local_arrow_cdm <- cdm_from_files(save_path, cdm_tables = all_of(cdm_tables), as_data_frame = FALSE)
+  local_arrow_cdm <- cdm_from_files(save_path, as_data_frame = FALSE)
   expect_s3_class(local_arrow_cdm, "cdm_reference")
   expect_equal(collect(local_arrow_cdm$person), collect(cdm$person))
-  expect_output(validate_cdm(local_arrow_cdm))
+  expect_error(validate_cdm(local_arrow_cdm))
 
   DBI::dbDisconnect(con, shutdown = TRUE)
 })