+++ tests/testthat/test-dbConnectDetails.R
@@ -3,8 +3,14 @@ library(dplyr, warn.conflicts = FALSE)
 library(DBI)
 
 self_contained_query <- function(connection_details, cdm_schema) {
-  con <- DBI::dbConnect(connection_details)
-  on.exit(DBI::dbDisconnect(con, shutdown = TRUE))
+  suppressMessages({
+    con <- DBI::dbConnect(connection_details)
+  })
+  if(dbms(con) == "duckdb") {
+    on.exit(DBI::dbDisconnect(con, shutdown = TRUE))
+  } else {
+    on.exit(DBI::dbDisconnect(con))
+  }
   DBI::dbGetQuery(con, paste0("select count(*) as n from ", cdm_schema, ".person"))
 }
 
@@ -52,7 +58,7 @@ test_that("dbConnectDetails works on sql server", {
   expect_true("n" %in% names(result))
 })
 
-test_that("dbConnectDetails works on sql server", {
+test_that("dbConnectDetails works on Redshift", {
   skip_if(Sys.getenv("CDM5_REDSHIFT_USER") == "")
 
   details <- dbConnectDetails(RPostgres::Redshift(),
@@ -67,7 +73,6 @@ test_that("dbConnectDetails works on sql server", {
   expect_true("n" %in% names(result))
 })
 
-
 test_that("dbConnectDetails works on duckdb", {
 
   details <- dbConnectDetails(duckdb::duckdb(), dbdir = eunomia_dir())
@@ -76,3 +81,65 @@ test_that("dbConnectDetails works on duckdb", {
   expect_s3_class(result, "data.frame")
   expect_true("n" %in% names(result))
 })
+
+# DatabaseConnector tests ----
+test_that("dbConnectDetails works on postgres using DatabaseConnector", {
+  skip_on_ci()
+  skip_if(Sys.getenv("CDM5_POSTGRESQL_USER") == "")
+
+  if (Sys.getenv("DATABASECONNECTOR_JAR_FOLDER") == "") {
+    Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = "~/DATABASECONNECTOR_JAR_FOLDER")
+    DatabaseConnector::downloadJdbcDrivers("postgresql")
+  }
+
+  details <- dbConnectDetails(DatabaseConnector::DatabaseConnectorDriver(),
+                              dbms = "postgresql",
+                              server = Sys.getenv("CDM5_POSTGRESQL_SERVER"),
+                              user = Sys.getenv("CDM5_POSTGRESQL_USER"),
+                              password = Sys.getenv("CDM5_POSTGRESQL_PASSWORD"))
+
+  result <- self_contained_query(details, Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"))
+  expect_s3_class(result, "data.frame")
+  expect_true("n" %in% names(result))
+})
+
+test_that("dbConnectDetails works on SQL Server using DatabaseConnector", {
+  skip_on_ci()
+  skip_if(Sys.getenv("CDM5_SQL_SERVER_USER") == "")
+
+  if (Sys.getenv("DATABASECONNECTOR_JAR_FOLDER") == "") {
+    Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = "~/DATABASECONNECTOR_JAR_FOLDER")
+    DatabaseConnector::downloadJdbcDrivers("sql server")
+  }
+
+  details <- dbConnectDetails(DatabaseConnector::DatabaseConnectorDriver(),
+                              dbms = "sql server",
+                              server   = Sys.getenv("CDM5_SQL_SERVER_SERVER"),
+                              user     = Sys.getenv("CDM5_SQL_SERVER_USER"),
+                              password = Sys.getenv("CDM5_SQL_SERVER_PASSWORD"))
+
+  result <- self_contained_query(details, Sys.getenv("CDM5_SQL_SERVER_CDM_SCHEMA"))
+  expect_s3_class(result, "data.frame")
+  expect_true("n" %in% names(result))
+})
+
+test_that("dbConnectDetails works on redshift using DatabaseConnector", {
+  skip_on_ci()
+  skip_if(Sys.getenv("CDM5_REDSHIFT_USER") == "")
+
+  if (Sys.getenv("DATABASECONNECTOR_JAR_FOLDER") == "") {
+    Sys.setenv("DATABASECONNECTOR_JAR_FOLDER" = "~/DATABASECONNECTOR_JAR_FOLDER")
+    DatabaseConnector::downloadJdbcDrivers("redshift")
+  }
+
+  details <- dbConnectDetails(DatabaseConnector::DatabaseConnectorDriver(),
+                              dbms = "redshift",
+                              server   = Sys.getenv("CDM5_REDSHIFT_SERVER"),
+                              user     = Sys.getenv("CDM5_REDSHIFT_USER"),
+                              password = Sys.getenv("CDM5_REDSHIFT_PASSWORD"))
+
+  result <- self_contained_query(details, Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"))
+  expect_s3_class(result, "data.frame")
+  expect_true("n" %in% names(result))
+})
+