+++ NEWS.md
@@ -1,3 +1,12 @@
+# CDMConnector 0.2.0
+
+* Add support for OMOP v5.3 and v5.4
+* Add datediff and dateadd functions that can be used in dplyr::mutate() 
+* Add assert_tables function for checking that a cdm object has required tables
+* Add camelCase versions of several functions
+
+Released Nov 15, 2022
+
 # CDMConnector 0.1.0
 
 * Initial release Sept 11, 2022